# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ioBroker Mobile Control API
  version: "1.0"
  description: |
    REST contract of the `iobroker.mobile-control` adapter, matching `docs/API-VERTRAG.md`
    and the implementation in `src/api/router.ts`. The adapter never exposes the raw
    ioBroker object tree or arbitrary state ids - only public object UUIDs from the
    filtered catalog. WebSocket contract (`/ws/v1`) is documented separately in
    `docs/API-VERTRAG.md` §8, since OpenAPI does not model WebSocket message flows.

    **Every `bearerAuth`-protected endpoint additionally requires a per-request signature**
    (see `src/security/requestSignature.ts`), on top of the bearer token - not modeled as a
    second OpenAPI security scheme since it isn't a simple header value but a computed
    signature over the request. Required headers:
    - `X-Signature-Timestamp`: current epoch milliseconds
    - `X-Signature-Nonce`: random, single-use per request
    - `X-Signature`: base64 DER-encoded ECDSA/SHA-256 signature (device Keystore key) over
      `METHOD\nPATH\nTIMESTAMP\nNONCE\nSHA256(body)`, where `PATH` starts at (and includes)
      `/api/v1` regardless of any external reverse-proxy prefix, and the body hash is of the
      exact raw request bytes (empty body -> hash of zero bytes). Missing or invalid headers
      return `SIGNATURE_INVALID` (401); a reused nonce returns `REPLAY_DETECTED` (409).
servers:
  - url: /api/v1
...
paths:
...
  /pairing/claim:
    post:
      tags: [pairing]
      summary: Claim a scanned pairing invitation and register this device
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimRequest'
      responses:
        '200':
          description: Claim accepted (still needs status polling)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimResponse'
        '400':
          $ref: '#/components/responses/Error'
  /pairing/status/{claimId}:
    get:
      tags: [pairing]
      summary: Poll the approval status of a pairing claim
      security: []
      parameters:
        - name: claimId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: |
            Status of the claim. Tokens are included exactly once - the first time this
            endpoint is polled after the claim has been approved. Subsequent polls only
            return `{"status":"approved"}` without tokens.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PairingStatusResponse'
        '404':
          $ref: '#/components/responses/Error'
  /auth/challenge:
    post:
      tags: [auth]
      summary: Request a signing challenge for an approved device
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deviceId]
              properties:
                deviceId: { type: string }
      responses:
        '200':
          description: Challenge issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeResponse'
        '403':
          $ref: '#/components/responses/Error'
  /auth/login:
    post:
      tags: [auth]
      summary: Exchange a signed challenge for access/refresh tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deviceId, challengeId, signature]
              properties:
                deviceId: { type: string }
                challengeId: { type: string }
                signature:
                  type: string
                  description: base64, DER-encoded ECDSA/SHA-256 signature over the raw challenge nonce bytes
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Error'
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate the refresh token and obtain a new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deviceId, refreshToken]
              properties:
                deviceId: { type: string }
                refreshToken: { type: string }
      responses:
        '200':
          description: New token pair. Reusing an already-rotated-away refresh token
            revokes the entire session/token family (see docs/SECURITY.md).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Error'
  /catalog:
    get:
      tags: [catalog]
      summary: Effective, filtered object catalog for the authenticated user/device
      security: [{ bearerAuth: [] }]
      parameters:
        - name: version
          in: query
          description: |
            Optional delta support: pass the `version` from a previously fetched catalog. If it
            still matches, the server responds with `{version, unchanged: true}` instead of the
            full object list, saving bandwidth. Omit to always get the full catalog.
          schema:
            type: integer
      responses:
        '200':
          description: "Full catalog, or a short unchanged-marker object if the passed version still matches"
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CatalogResponse'
                  - type: object
                    properties:
                      version: { type: integer }
                      unchanged: { type: boolean, enum: [true] }
        '401':
          $ref: '#/components/responses/Error'
  /states:
    get:
      tags: [catalog]
      summary: Current values of one or more public object UUIDs
      security: [{ bearerAuth: [] }]
      parameters:
        - name: ids
          in: query
          required: true
          description: Comma-separated list of public object UUIDs
          schema:
            type: string
      responses:
        '200':
          description: |
            Values keyed by the requested UUID. A key maps to `null` when the object can't be
            resolved for this caller (unknown/unauthorized) - callers must not assume every
            requested id is present with a non-null value.
          content:
            application/json:
              schema:
                type: object
                properties:
                  states:
                    type: object
                    additionalProperties:
                      nullable: true
                      $ref: '#/components/schemas/StateValue'
        '401':
          $ref: '#/components/responses/Error'
  /history:
    get:
      tags: [catalog]
      summary: Historical values of a single public object UUID
      description: |
        Only available for objects whose exposure rule has `history: true`, and only if the
        adapter is configured with a history-storing instance (see docs/DEPLOYMENT.md). Returns
        an empty list rather than an error when no history instance is configured.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: query
          required: true
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          schema: { type: integer, default: 500, maximum: 2000 }
      responses:
        '200':
          description: History entries, oldest-to-newest order as returned by the history adapter
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      type: object
                      properties:
                        value: {}
                        timestamp: { type: string, format: date-time }
        '403':
          description: READ_FORBIDDEN - object not visible, or history not enabled for it
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          $ref: '#/components/responses/Error'
  /objects/{id}/snapshot:
    get:
      tags: [catalog]
      summary: Camera snapshot image for a public object UUID
      description: |
        Proxies the current snapshot of a camera-backed ioBroker state so the app never needs
        direct network access to the camera or the adapter serving it. The underlying state must
        hold either a data URL ("data:image/jpeg;base64,...") or a plain http(s) URL to a snapshot
        image - any other value shape returns SERVER_UNAVAILABLE. See src/camera/index.ts.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: The snapshot image bytes, Content-Type set from the source; X-Snapshot-Timestamp header carries the ioBroker state timestamp
          content:
            image/*:
              schema: { type: string, format: binary }
        '403':
          description: READ_FORBIDDEN - object not visible to this caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: SERVER_UNAVAILABLE - no snapshot value, unsupported value shape, or the source URL did not respond
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          $ref: '#/components/responses/Error'
  /url-embeds:
    get:
      tags: [url-embeds]
      summary: List admin-approved URL embeds available to this device
      description: |
        Admin-managed allowlist (see src/urlEmbeds/index.ts) - deliberately returns only {id, name}
        for every entry, never the underlying URL. A client cannot fetch or navigate to a URL of
        its own choosing; it can only reference an id from this list via the two endpoints below.
        Only returns embeds this role/user/device has been explicitly granted access to (same
        default-deny, deny-wins-outright priority as object exposure rules) - an embed with no
        access rule at all is invisible here even though it exists in the admin's allowlist.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: The list of allowlisted embeds
          content:
            application/json:
              schema:
                type: object
                properties:
                  embeds:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        name: { type: string }
        '401':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
...
    ClaimRequest:
      type: object
      required: [pairingId, pairingSecret, deviceName, platform, appVersion, publicKey]
      properties:
        pairingId: { type: string }
        pairingSecret: { type: string }
        deviceName: { type: string }
        platform: { type: string, example: android }
        appVersion: { type: string }
        publicKey:
          type: string
          description: base64 SPKI-encoded EC P-256 public key
    ClaimResponse:
      type: object
      properties:
        claimId: { type: string }
        status:
          type: string
          enum: [waiting_for_approval, approved]
    PairingStatusResponse:
      type: object
      properties:
        status:
          type: string
          enum: [waiting_for_approval, approved, rejected, expired]
        deviceId: { type: string, nullable: true }
        accessToken: { type: string, nullable: true }
        refreshToken: { type: string, nullable: true }
        expiresIn: { type: integer, nullable: true, description: seconds }
...